requirements.txt
langchain-groq
langchain-openai
pydantic
google-genai
langgraph
requests
fastapi
uvicorn 
python-dotenv
streamlit